Init new version

+++ installer.sh
@@ -17,8 +17,10 @@
 #wget https://preview.c9users.io/romanrehacek/installers/x.sh -v -O install.sh && chmod +x install.sh && ./install.sh; rm -rf install.sh
 
 echo "\033[44mDeleting unnecessary files and folders\033[m\n"
-if [ -f README.md ]; then
-    rm README.md
+if [ ! -d ".c9" ]; then
+    if [ -f README.md ]; then
+        rm README.md
+    fi
 fi
 
 if [ -f php.ini ]; then
@@ -82,30 +84,91 @@ if [ -d "wp-includes" ]; then
         
         if test "$pathToTheme" != ""; then
             rm -R -- ${pathToTheme}* 
+            
+            echo "\033[44mDownload default theme...\033[m\n"
             wget "https://github.com/romanrehacek/default-wp-theme/archive/master.zip" -N -q
             if [ ! -f master.zip ]; then
                 echo "\n\033[41mERROR: master.zip didn't downloaded!\033[m\n"
                 exit
             fi
-            unzip master.zip -d $pathToTheme
+            
+            echo "\033[44mExtract default theme...\033[m\n"
+            unzip -q master.zip -d $pathToTheme
             rm master.zip
             mv ${pathToTheme}default-wp-theme-master ${pathToTheme}${dirName}
             
             echo -n "\n\033[41mEnter theme name\033[m (leave blank for 'Example theme'): "
             read themeName
+            themeName=${themeName:-'Example theme'}
             
             echo -n "\n\033[41mEnter site domain\033[m (leave blank for 'example.com'): "
             read domainName
+            domainName=${domainName:-'example.com'}
             
             sed -i "s/\[theme_name\]/$(echo $themeName | sed -e 's/[\/&]/\\&/g')/g" ${pathToTheme}${dirName}/style.css
             sed -i "s/\[domain_name\]/$(echo $domainName | sed -e 's/[\/&]/\\&/g')/g" ${pathToTheme}${dirName}/style.css
             sed -i "s/\[dir_name\]/$(echo $dirName | sed -e 's/[\/&]/\\&/g')/g" ${pathToTheme}${dirName}/style.css
+            
+            sed -i "s/theme_name/$(echo $dirName | sed -e 's/[\/&]/\\&/g')/g" $(find ${pathToTheme}${dirName}/ -type f -not -iwholename '*.svn*' -not -iwholename '*.git*' -not -iwholename '*.c9*')
+            sed -i "s/Theme_name/$(echo $dirName | sed 's/.*/\u&/' | sed -e 's/[\/&]/\\&/g')/g" $(find ${pathToTheme}${dirName}/ -type f -not -iwholename '*.svn*' -not -iwholename '*.git*' -not -iwholename '*.c9*')
+            sed -i "s/THEME_NAME/$(echo $dirName | awk '{print toupper($0)}' | sed -e 's/[\/&]/\\&/g')/g" $(find ${pathToTheme}${dirName}/ -type f -not -iwholename '*.svn*' -not -iwholename '*.git*' -not -iwholename '*.c9*')
+            
+            find ${pathToTheme}${dirName}/ -type f -not -iwholename '*.svn*' -not -iwholename '*.git*' -not -iwholename '*.c9*' -name "*theme_name*" | sed -e "p;s/theme_name/$(echo $dirName | sed -e 's/[\/&]/\\&/g')/g" | xargs -n2 mv
         fi
     fi
 fi
 
-if [ -f start.sh ]; then
-    rm start.sh
+if ([ -d "wp-includes" ] && test "$new_theme" = "n" ) || ( [ ! -d "wp-includes" ] ); then
+    echo -n "\n\033[41mInstall main files? [y/n]:\033[m "
+    read main_files
+    
+    if test "$main_files" = "y"; then
+        dirName=$(basename "$path")
+        
+        pathToTheme=$(echo $path | sed 's!'"$dirName"'.*!!g')
+        
+        if test "$pathToTheme" = ""; then
+            pathToTheme=$(echo $path)
+        fi
+        
+        if test "$pathToTheme" != ""; then
+            
+            echo "\033[44mDownload default theme...\033[m\n"
+            wget "https://github.com/romanrehacek/default-wp-theme/archive/master.zip" -N -q
+            if [ ! -f master.zip ]; then
+                echo "\n\033[41mERROR: master.zip didn't downloaded!\033[m\n"
+                exit
+            fi
+            
+            if [ ! -d "$path" ]; then
+                echo "\033[44mCreate directory $path...\033[m\n"
+                mkdir -p $path;
+            fi
+            
+            echo "\033[44mExtract default theme...\033[m\n"
+            unzip -q master.zip -d $pathToTheme
+            rm master.zip
+            
+            mkdir -p ${pathToTheme}default-wp-theme-master/assets/* ${path}assets/
+            mv -v ${pathToTheme}default-wp-theme-master/assets/* ${path}assets/
+            rm -rf ${pathToTheme}default-wp-theme-master
+            
+            echo -n "\n\033[41mEnter directory name\033[m (leave blank for 'xxx'): "
+            read themeName
+            
+            # Set default value
+            themeName=${themeName:-xxx}
+            
+            sed -i "s/THEME_NAME/$(echo $themeName | awk '{print toupper($0)}' | sed -e 's/[\/&]/\\&/g')/g" $(find ${path}assets/ -type f -not -iwholename '*.svn*' -not -iwholename '*.git*' -not -iwholename '*.c9*' -not -iwholename '*installer*' -not -iwholename '*gitignore*' -not -iwholename '*package.json*' -not -iwholename '*gulpfile.js*' -not -iwholename '*node_modules*')
+            
+        fi
+    fi
+fi
+
+if [ ! -d ".c9" ]; then
+    if [ -f start.sh ]; then
+        rm start.sh
+    fi
 fi
 
 echo "\033[44mInstall Git WebUI...\033[m\n"
